fix(초대 메일): 시각을 「어느 나라 시간」으로 보이게 + 일정 파일 첨부
문제(PO 제보 2026-08-03): 초대 메일 시각이 "15:00 GMT+9 · 06:00 UTC" 로만
적혀 있어 ①어느 나라 시간인지 안 보이고 ②받는 사람(카자흐·러시아)이 직접
환산해야 했다. 숫자 자체는 맞았지만 노쇼로 이어지기 쉬운 표기였다.

바꾼 것
- 한국 시각을 나라 이름으로: "GMT+9" → "대한민국 표준시 / Корея, стандартное
  время / Korean Standard Time" (Intl 의 timeZoneName:"long")
- 일정 파일(.ics) 첨부: 받는 사람 달력이 각자 시간대로 그려주고 30분 전 알람도
  걸린다. 코디가 상대 국적을 고를 필요가 없어 사람 실수가 낄 자리가 없다.
- 첨부가 안 열릴 때의 대비로 주요 도시 시각 한 줄(러/카자흐 메일: 알마티·
  타슈켄트·비슈케크·모스크바, 중국어: 北京, 일본어: 東京). 목록은 «메일 언어»로
  자동 결정 — 화면에 고르는 칸을 만들지 않았다.
- 시각 표시는 한 줄 나열 대신 굵은 한국 시각 + 회색 UTC 로 쌓았다(긴 시간대
  이름에서 줄이 흘러넘쳐 라벨과 어긋나던 것 교정).
- sendEmail 에 첨부 지원(Resend 경로). SES 로 떨어지면 첨부가 빠지므로 경고 로그.

검증
- 자동 검사 7건: 시각 문구(러/한), 도시 시각 계산값, 일정 파일 형식(UTC 저장·
  CRLF·이스케이프)
- 실발송 1통으로 admin@healwith.co.kr 수신 확인 — 첨부 healwith-consultation.ics
  (551바이트)가 달력 파일로 인식됨, 본문 3줄 정상

+++ src/lib/email/templates/consultationReminder.ts
@@ -133,7 +133,9 @@ export function renderConsultationReminderEmail(props: ConsultationReminderProps
       hour: "2-digit",
       minute: "2-digit",
       timeZone: "Asia/Seoul", // 상담은 한국시간 진행 — 서버(UTC) 기준으로 찍히면 시각이 틀림
-      timeZoneName: "short",
+      // "GMT+9" 만으론 어느 나라 시간인지 안 보인다 → long("한국 표준시"). 상대 현지 시각은
+      // 초대 메일에 첨부한 일정 파일이 달력에서 자동으로 맞춰 준다(사람이 국가를 고르지 않는다).
+      timeZoneName: "long",
     }
   );
